refactor: 1:1 대응 hook 위치 이동
